+++ src/components/notice/form/NoticeForm.tsx
@@ -1,28 +1,21 @@
-import { useCallback, useEffect } from 'react';
-
 import { useNavigate } from 'react-router-dom';
+
+import { useHandlePostNotice } from '@/hooks/useHandlePostNotice';
 
 import {
   initialUserProfile,
   useGetUserProfile,
 } from '@/services/auth/authQueries';
-import { usePostNotice } from '@/services/notice/noticeMutation';
 
 import {
   defaultNoticeFormValues,
   noticeCategoryOptions,
   specialLectureEventFormValues,
 } from '@/constants';
-import { useDialogContext, usePageBlocker } from '@/hooks';
+import { usePageBlocker } from '@/hooks';
 import { NoticeDto, SpecialLectureOrEventValue } from '@/types';
 import { zodResolver } from '@hookform/resolvers/zod';
-import {
-  Controller,
-  FormProvider,
-  SubmitErrorHandler,
-  useForm,
-  useWatch,
-} from 'react-hook-form';
+import { Controller, FormProvider, useForm, useWatch } from 'react-hook-form';
 
 import CircleNumber from '@/components/common/CircleNumber';
 import Title from '@/components/common/Title';
@@ -31,7 +24,6 @@ import MultiSelectDropdown from '@/components/common/dropdown/MultiSelectDropdow
 import SingleSelectDropdown from '@/components/common/dropdown/SingleSelectDropdown';
 import LabeledSection from '@/components/common/input/LabeledSection';
 import ControllerContentEditor from '@/components/common/text-editor/ControllerContentEditor';
-import { Session } from '@/components/notice/form/ControllerSessions';
 import ExtraInfoForm from '@/components/notice/form/ExtraInfoForm';
 import {
   NoticeCategoryKeySchemaType,
@@ -54,119 +46,14 @@ export default function NoticeForm() {
 
   const { data: userProfile = initialUserProfile } = useGetUserProfile();
 
-  const { hideDialog, alert, showToast } = useDialogContext();
+  const { onSubmit, onError, findCurrNotice } = useHandlePostNotice();
 
   const navigate = useNavigate();
 
-  const { blocker } = usePageBlocker({
+  usePageBlocker({
     isBlockRefresh: true,
     form: { isDirty, isSubmitted },
   });
-
-  useEffect(() => {
-    if (blocker.state === 'blocked') {
-      alert({
-        text: '정말 나가시겠어요?',
-        subText: '저장하지 않은 내용을 잃어버릴 수 있어요.',
-        children: (
-          <>
-            <SquareButton
-              color="gray"
-              name="계속 작성하기"
-              onClick={() => {
-                hideDialog();
-                if (blocker.state === 'blocked') {
-                  blocker.reset();
-                }
-              }}
-            />
-            <SquareButton
-              name="나가기"
-              onClick={() => {
-                hideDialog();
-                if (blocker.state === 'blocked') {
-                  blocker.proceed();
-                }
-              }}
-            />
-          </>
-        ),
-      });
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [blocker.state]);
-
-  const { mutate } = usePostNotice({
-    onError: () => {
-      alert({
-        text: '오류가 발생했습니다.',
-        subText: '다시 시도해주세요.',
-        children: (
-          <SquareButton
-            name="확인"
-            onClick={() => {
-              hideDialog();
-              navigate('/notice');
-            }}
-          />
-        ),
-      });
-    },
-    onSuccess: () => {
-      alert({
-        text: '공지사항이 등록되었습니다!',
-        children: (
-          <SquareButton
-            name="확인"
-            onClick={() => {
-              hideDialog();
-              navigate('/notice');
-            }}
-          />
-        ),
-      });
-    },
-  });
-
-  const onError: SubmitErrorHandler<NoticeDto.Post> = errors => {
-    const firstErrorKey = Object?.keys(errors)?.[0] as keyof NoticeDto.Post;
-    const firstErrorMsg = errors[firstErrorKey]?.message;
-    if (firstErrorMsg) {
-      showToast(firstErrorMsg);
-    }
-  };
-
-  const noticeKey = useWatch({ control, name: 'noticeType' });
-
-  const findCurrNotice = useCallback((key: NoticeCategoryKeySchemaType) => {
-    return noticeCategoryOptions.find(option => key === option.key);
-  }, []);
-
-  const onSubmit = async (submittedValue: NoticeDto.Post) => {
-    const {
-      noticeType,
-      title,
-      content,
-      targetCourseIdList,
-      satisfactionSurvey,
-      sessions,
-    } = submittedValue;
-
-    const sessionsWithNoId = (sessions as Session[])?.map(
-      ({ id, ...rest }) => rest,
-    );
-    if (noticeType === 'SPECIAL_LECTURE' || noticeType === 'EVENT') {
-      const formValue = {
-        ...submittedValue,
-        satisfactionSurvey: satisfactionSurvey || '',
-        sessions: sessionsWithNoId,
-      };
-      mutate(formValue);
-    } else {
-      const formValue = { noticeType, title, content, targetCourseIdList };
-      mutate(formValue);
-    }
-  };
 
   const setConditionalKey = (type: NoticeCategoryKeySchemaType) => {
     if (type === 'SPECIAL_LECTURE' || type === 'EVENT') {
@@ -176,6 +63,8 @@ export default function NoticeForm() {
       reset({ noticeType, title, content, targetCourseIdList });
     }
   };
+
+  const noticeKey = useWatch({ control, name: 'noticeType' });
 
   return (
     <FormProvider {...methods}>
